+++ Products/RedirectionTool/permissions.py
@@ -1,1 +1,2 @@
+# -*- coding: utf-8 -*-
 ModifyAliases = "RedirectionTool: Modify aliases"